DHCP support

+++ clock-arch.h
@@ -34,7 +34,7 @@
 #ifndef __CLOCK_ARCH_H__
 #define __CLOCK_ARCH_H__
 
-typedef int clock_time_t;
+typedef unsigned long clock_time_t;
 #define CLOCK_CONF_SECOND 1000
 
 #endif /* __CLOCK_ARCH_H__ */
